fix: 3 build errors - log.h forward decl, AddRectFilledMultiColorRounded, int64 cast
- log.h: move refresh() after render() so render() is declared first\n- modern_ui.cpp: AddRectFilledMultiColor -> AddRectFilledMultiColorRounded (7th arg was rounding)\n- cache.cpp: static_cast<uint8_t> for rigType assignment to suppress C4244

+++ src/core/cache.cpp
@@ -282,7 +282,7 @@ namespace cache {
                     player.*(it->second) = inst;
                 }
             }
-            player.Rig_Type = cached->rigType;
+            player.Rig_Type = static_cast<uint8_t>(cached->rigType);
         }
 
         // ---- Volatile data (updated every tick) ----
